Added GET requests for: AdventurerJobs, JobPage, LocationPage, RankPagee, AdventurerPage.
Added a POST request with a button in HomePage for a database reset.

backend/server contains all of the get and post requests.

Added a PL.SQL file for the database reset.

Updated the Step3_DML file for the adventurerJobPage for syntax.

Added a refresh button for the AdventureJobsPage to reset the page if entering from the Adventure page.